% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExprEstimators.R
\name{estimateMAValues}
\alias{estimateMAValues}
\title{Estimation of the M and A Values}
\usage{
estimateMAValues(R.mean, G.mean, R.var, G.var, RG.cov, R.bckg, G.bckg,
  estimator = "second-order", bgcorr = "normexp", offset = 50,
  putaway.perc = 0.05, array.ids = NULL, gene.ids = NULL)
}
\arguments{
\item{R.mean}{a p x n matrix of mean values for the red (test) foreground intensities.}

\item{G.mean}{a p x n matrix of mean values for the green (reference) foreground intensities.}

\item{R.var}{a p x n matrix of variances of the red foreground intensities.}

\item{G.var}{a p x n matrix of variances of the green foreground intensities.}

\item{RG.cov}{a p x n matrix of covariances between the green and red foreground intensities.}

\item{R.bckg}{a p x n matrix of mean values of the red background intensities.}

\item{G.bckg}{a p x n matrix of mean values of the green background intensities.}

\item{estimator}{a character string indicating the estimation method. The options are 'conventional' and 'second-order'.}

\item{bgcorr}{a character string indicating the background correction method. The options are 'none' and 'normexp'.}

\item{offset}{a numeric parameter for the normexp background correction.}

\item{putaway.perc}{a scalar between 0 and 100 indicating percentage of genes expressions.
with high variance to be discarded. Default is 0.05.}

\item{array.ids}{a character vector with identifiers of the arrays.}

\item{gene.ids}{a character vector with identifiers of the genes.}
}
\value{
a limma's MAlist object with the M and A values
}
\description{
Method for estimating the intensity values M and A, considering p spots/genes and n microarray slides.
 If estimator is set as 'second-order', then an estimator that takes into account pixel-level inaccuracies is used.
}
\examples{
\donttest{
data(metaplasia)
MA <- estimateMAValues(metaplasia$R.mean, metaplasia$G.mean,
                       metaplasia$R.var, metaplasia$G.var, metaplasia$RG.cov,
                       metaplasia$R.bckg, metaplasia$G.bckg,
                       array.ids=metaplasia$array.ids, gene.ids=metaplasia$gene.ids)
}

}
