% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExprEstimators.R
\name{getOptimalLowessFvalue}
\alias{getOptimalLowessFvalue}
\title{Optimal Selection of the LOWESS Bandwidth Parameter Using the HRCP Criterion}
\usage{
getOptimalLowessFvalue(M.mean, A.mean, eva.values = seq(0.2, 1, by =
  0.05), debug = TRUE, save.objs = TRUE, save.plots = TRUE,
  dir.to.save = "./", array.id = "array")
}
\arguments{
\item{M.mean}{a vector with the M values for the p genes.}

\item{A.mean}{a vector with the A values for the p genes.}

\item{eva.values}{a vector with bandwith values to be considered by the LOWESS parameter selection method.}

\item{debug}{Logical indication if you want to view logs of the execution.}

\item{save.objs}{Logical indicating if you want to save objects with partial results.}

\item{save.plots}{Logical indicating if you want to generate the M plots
illustrating the bandwidth parameter selection process.}

\item{dir.to.save}{path to the folder you want to save the output objects.}

\item{array.id}{a character string identifying the array.}
}
\value{
a MAList object, as in the limma R package,
with the MA values after applying within-slide normalization by LOWESS with optimal parameter settings.
}
\description{
Data-driven parameter selection method that is parsimonious and considers
intrinsic characteristics of microarray data, such as heteroskedasticity. Particularly, the
best bandwidth is selected according to the HRCp criterion.
}
\references{
{
  \insertRef{liu2013heteroscedasticity}{omicsMA}
}
}
