% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metaplasia}
\alias{metaplasia}
\title{Intestinal Metaplasia Dataset}
\format{A list with the following elements:
\describe{
\item{R.mean}{A 45015 x 10 matrix of mean values for the red (test) foreground intensities.}
\item{G.mean}{A 45015 x 10 matrix of mean values for the green (reference) foreground intensities.}
\item{R.var}{A 45015 x 10 matrix of variances of the red foreground intensities.}
\item{G.var}{A 45015 x 10 matrix of variances of the green foreground intensities.}
\item{RG.cov}{A 45015 x 10 matrix of covariances between the green and red foreground intensities.}
\item{R.bckg}{A 45015 x 10 matrix of mean values of the red background intensities.}
\item{G.bckg}{A 45015 x 10 matrix of mean values of the green background intensities.}
\item{R.size}{A 45015 x 10 matrix with the number of red foreground pixels.}
\item{G.size}{A 45015 x 10 matrix with the number of green foreground pixels.}
\item{array.ids}{A vector with the ids for the 10 arrays.}
\item{gene.ids}{A vector with the ids for the 45015 spots.}
\item{gene.names}{A vector with the names of the 45015 genes.}
}}
\usage{
metaplasia
}
\description{
This is part of dataset used in the study by Ribeiro A H et al. entitled 'Variance-Preserving
Estimation of Intensity Values Obtained fom Omics Experiments'.
}
\details{
It contains data from 10 two-color microarrays, being 5 from tissues representing
type II intestinal metaplasia and 5 from tissues representing the normal condition,
obtained from the Tumor Bank at A.C. Camargo Cancer Center / Antonio Prudente Foundation.

Each sample is hybridized against a pool of normal tissues using the same orientation of
dye labeling. Gene expression levels were measured on
Agilent Whole Human Genome Microarrays 4x44K G4112F (design ID 014850), each slide
containing 45015 spots (41093 unique probes).

The scanned images of the microarray slides were processed
by Agilent Feature Extraction software, version 9.5, where statistics
(mean, standard deviation and covariance) of the foreground and local background pixels were
computed for each spot, in both test and reference channels.
Each microarray spot contains about 60 foreground pixels.
}
\keyword{datasets}
