% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExprEstimators.R
\name{normalizeWithinArraysByOptimalLowess}
\alias{normalizeWithinArraysByOptimalLowess}
\title{Optimal LOWESS Within-Slide Normalization of a MAList}
\usage{
normalizeWithinArraysByOptimalLowess(MA, eva.values = seq(0.2, 1, by =
  0.05), debug = TRUE, save.objs = TRUE, save.plots = TRUE,
  dir.to.save = "./")
}
\arguments{
\item{MA}{a MAList object, as in the limma R package, with the non-normalized MA values.}

\item{eva.values}{a vector with bandwith values to be considered by the LOWESS parameter selection method.}

\item{debug}{Logical indication if you want to view logs of the execution.}

\item{save.objs}{Logical indicating if you want to save objects with partial results.}

\item{save.plots}{Logical indicating if you want to generate the M plots
illustrating the bandwidth parameter selection process.}

\item{dir.to.save}{path to the folder you want to save the output objects.}
}
\value{
a MAList object, as in the limma R package,
with the MA values after applying within-slide normalization by LOWESS with optimal parameter settings.
}
\description{
Applies LOWESS for within-slide normalization of the MA values.
The parameters are set by a data-driven parameter selection method that is parsimonious and considers
intrinsic characteristics of microarray data, such as heteroskedasticity. Particularly, the
best bandwidth is selected according to the HRCp criterion.
}
\examples{
\donttest{
data(metaplasia)
MA <- estimateMAValues(metaplasia$R.mean, metaplasia$G.mean,
                       metaplasia$R.var, metaplasia$G.var, metaplasia$RG.cov,
                       metaplasia$R.bckg, metaplasia$G.bckg,
                       array.ids=metaplasia$array.ids, gene.ids=metaplasia$gene.ids)
normMA <- normalizeWithinArraysByOptimalLowess(MA)
}

}
\references{
{
  \insertRef{liu2013heteroscedasticity}{omicsMA}
}
}
